task 1
This C++ program is designed to engage the user in a simple number-guessing game. The program will generate a random number between 1 and 100 and prompt the user to guess this number. Based on the user's input, the program will provide feedback indicating whether the guess is too high or too low. This process will continue until the user successfully guesses the correct number, at which point the program will congratulate the user.